backend/assignments/urls.py
"""
URLs for assignments app.
"""
from django.urls import path, include
from rest_framework.routers import DefaultRouter
from .views import AssignmentViewSet

router = DefaultRouter()
router.register(r'', AssignmentViewSet, basename='assignment')

urlpatterns = [
    path('', include(router.urls)),
]
